Added code to handle duplicate name cases.
If user tries to add the same contact name with different case it
will not be allowed.

If user tries to add a similar contact name (e.g with more spaces),
the new contact will be added but a warning will be shown to inform
them that a similar contact name already exists.